At the user levl, we wait for a sequence. We need to add a MORSE_Barrier function that waits for all sequences currently running

+++ example/lapack_to_morse/step3.c
@@ -36,8 +36,8 @@ int main(int argc, char *argv[]) {
 
     size_t i, j;
     size_t N;    // matrix order
-    size_t NB;   // number of rows and columns in tiles
-    size_t NRHS; // number of RHS vectors
+    int NB;   // number of rows and columns in tiles
+    int NRHS; // number of RHS vectors
     int NCPU; // number of cores to use
     int NGPU; // number of gpus (cuda devices) to use
     int UPLO = MorseUpper; // where is stored L
